app/code/requirements.txt
accelerate
altgraph
annotated-doc
annotated-types
anyio
certifi
charset-normalizer
click
colorama
distro
exceptiongroup
fastapi
filelock
fsspec
h11
hf-xet
httpcore
httpx
huggingface-hub
idna
Jinja2
jiter
MarkupSafe
modelscope
mpmath
networkx
numpy
openai
opencv-python
packaging
pefile
pillow
psutil
pydantic
pydantic_core
pyinstaller
pyinstaller-hooks-contrib
PyMuPDF
pywin32-ctypes
PyYAML
regex
requests
safetensors
sniffio
starlette
sympy
tiktoken
timm
tokenizers
torch
torchvision
tqdm
transformers
typing-inspection
typing_extensions
urllib3
uvicorn